pyproject.toml
[tool.poetry]
name = "n26-pdf-converter"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "pdf-converter"}]

[tool.poetry.dependencies]
python = "^3.12"
pandas = {extras = ["pyarrow"], version = "^2.2.0"}
tabula-py = "^2.9.0"
typer = {extras = ["all"], version = "^0.9.0"}
pyarrow = "^15.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
